Chore: update devcontainer config
Closes #13.

 - install Talisman binary as pre-commit build from source is broken
 - remove use of deprecated apt-key command
 - install kubectl from GCP Cloud SDK
 - install Terraform from Hashicorp apt repo
 - install terraform-docs 0.15.0 from tarball
 - install pre-commit to non-standard hook dir and configure git to use
   it for hooks

VS Code automatically mounts the workspace as a volume in the container; this
means that when pre-commit is installed, the .git/hooks/pre-commit will be
configured to use global python from container, possibly breaking hook when
invoked outside of the container.

Since ~/.gitconfig is *copied* into the container, add a postAttach action to
set the global config in container to use non-standard hook path.